workers/layout_worker/requirements.txt
docling>=2.5.0
marker-pdf>=1.10.2
torch>=2.4.0
